+++ Products/CMFFormController/FSControllerPythonScript.py
@@ -31,15 +31,14 @@ from FSControllerBase import FSControllerBase
 from utils import log, logException
 from interfaces import IControllerPythonScript
 
-from zope.interface import implements
+from zope.interface import implementer
 
+@implementer(IControllerPythonScript)
 class FSControllerPythonScript (FSControllerBase, BaseClass):
     """FSControllerPythonScripts act like Controller Python Scripts but are not
     directly modifiable from the management interface."""
 
     meta_type = 'Filesystem Controller Python Script'
-
-    implements(IControllerPythonScript)
 
     manage_options=(
            (